CG-11: Moved constants.h to dreid.h to match new project name.
