+++ Makefile
@@ -2,16 +2,16 @@ CC = g++
 CFLAGS = -g -std=c++20
 INCLUDES=#-I /usr/local/BerkeleyDB.18.1/include # -I /usr/include/mysql-cppconn-8
 LIBS=-L/usr/lib/x86_64-linux-gnu -lpthread # -lmysqlcppconn8 -lmysqlcppconn
-HEADERS = config.h constants.h csvtools.h dht.h dq.h
+HEADERS = config.h dreid.h csvtools.h dht.h dq.h
 OBJECTS = board.o util.o gameinfo.o piece.o position.o worker.o csvtools.o md5.o dht.o dq.o
 
 .cpp.o:
 	$(CC) $(CFLAGS) $(INCLUDES) -c $<
 
-all: cg
+all: dr
 
-cg : chessboard.o $(OBJECTS) $(HEADERS)
-	$(CC) $(CFLAGS) chessboard.o $(OBJECTS) $(LIBS) -o cg
+dr : chessboard.o $(OBJECTS) $(HEADERS)
+	$(CC) $(CFLAGS) chessboard.o $(OBJECTS) $(LIBS) -o dr
 
 bang0 : bang0.o $(OBJECTS) $(HEADERS)
 	$(CC) $(CFLAGS) bang0.o $(OBJECTS) $(LIBS) -o bang0
@@ -31,7 +31,7 @@ clean:
 clean-dq :
 	rm -rf ~/tmp/testqueue
 
-clean-cg :
+clean-dr :
 	rm -rf ~/work/data
 
 test : test.o $(OBJECTS) $(HEADERS)
