+++ csvtools.h
@@ -7,7 +7,7 @@
 #include <sstream>
 #include <stdexcept>
 
-#include "constants.h"
+#include "dreid.h"
 
 int load_csv(std::string filename, PosMap& map);
 uint64_t csv_cmp(std::string filename, PosMap& lhs);
