+++ dreid.h
