+++ fenutil.cpp
@@ -6,7 +6,7 @@
 //
 #include <iostream>
 
-#include "constants.h"
+#include "dreid.h"
 
 void usage()
 {
