+++ worker.h
@@ -3,7 +3,7 @@
 #include <fstream>
 #include <string>
 
-#include "constants.h"
+#include "dreid.h"
 #include "dht.h"
 #include "dq.h"
